Switch to point to the UFS_UTILS repository  (#15)
Changes from the feature/test_ufs_utils branch:
Switch to point to the feature/HAFS branch of the UFS_UTILS repo (instead of the gfs-workflow repo) for chgres, fre-nctools, and orog. And the related changes from the HAFS branch in the gfs-workflow repo have been ported into the feature/HAFS branch of UFS_UTILS.
The fv3gfs dir is deleted since it is no longer needed.
The sorc/hafs_utils.fd now points to the feature/HAFS branch of the UFS_UTILS repo.
The original sorc/hafs_utils.fd was renamed to sorc/hafs_tools.fd.
Fix the build system after switching to point to the UFS_UTILS repo.
And update to point the latest changes for all the submodules.