Porting the existing HAFS python scripts from Python2 to Python3 (#42)
This Pull Request is to address issue #31 and merge the feature/HAFSpy3 branch back to the develop branch.

The majority of porting work from Python2 to Python3 is done by @SamuelTrahanNOAA and DTC during the HAFS-CROW connection. The porting for pyprodutils was mainly done by @SamuelTrahanNOAA and @jimfrimel. And @BinLiu-NOAA and @evankalina conducted some testing and further modifications to enable supporting multi-platforms and get the branch ready for the merge.

This feature/HAFSpy3 branch has been recently synced with the latest develop branch, and tested on Hera (producing bit-wisely identical results). The HAFS workflow regression tests also passed successfuly on other platforms including (Jet, wcoss_cray and wcoss_dell_p3).