Feature/hrd graphics (#63)
This Pull Request addresses the issue #62. Include the HRD GPLOT graphics in the HAFS workflow. There is a run_hrdgraphics config option to switch on and off the hrdgraphics task in the HAFS workflow. Currently, run_hrdgraphics=no by default.

Note: Only Jet is extensively tested. Support and testing on other platforms are needed.

Co-authored-by: ghassan-alaka <ghassan.alaka@noaa.gov>, BinLiu-NOAA, AndrewHazelton