Add data model support to HAFS workflow (#91)
This PR allows developers to perform HAFS runs with the Community Data models for Earth Prediction Systems (CDEPS). Developers can use the new workflow capabilities to one-way couple the HYCOM ocean model to a data atmosphere created from the ERA5 reanalysis. Or they can one-way couple the UFS weather model to a data ocean created from the OISST or GHRSST analyses. The workflow is capable of running all of the necessary steps in an automated fashion, including pre-processing the input data, generating the data model grids, and running the forecast. Scripts are provided to download ERA5, OISST, and GHRSST data, which must be run offline. Documentation is available for exercising CDEPS support within the HAFS workflow. This PR is not expected to have any impact on HAFS runs that do not use data models.

Co-authored-by: Samuel Trahan <Samuel.Trahan@noaa.gov>
Co-authored-by: Ufuk Turuncoglu <turuncu@ucar.edu>
Co-authored-by: Bin.Liu <Bin.Liu@noaa.gov>
Co-authored-by: john.steffen <john.steffen@noaa.gov>
Co-authored-by: Daniel Rosen <drosen@ucar.edu>